Added write_column_names function, accounted for some edge cases
Moved functionality write the three column titles to its own function. Changed the input arguments to only require worksheet and the movie data, as opening the worksheet is done before the call to the function.

Dealt with several edge cases like not allowing to save to the workbook while the file is still open and dealing with potential empty lines in the input file.