Finsihed a basic README file and Input Edits
The README describes the project from a user perspective along with my approach. The project is still simple in its current state, but the description includes some plans to improve the project and add more features.

I also made some minor changes to the example input files and added another useful link that I found while working on this project.